p2p: Use `FeeRate` in the feefilter message
The previous `i64` is obviously not type-safe and I believe it is also
wrong. BIP-133 states this value is encoded as a uint64_t. The `u64` is
decoded of the wire and interpreted as sats per kvb.

Given the `FeeRate` constructor accepts a `u32`, if the `try_into`
overflows we can treat it as `u32::MAX`, which essentially says the
remote node doesn't want any transactions from us.

ref: https://github.com/bitcoin/bips/blob/master/bip-0133.mediawiki#specification